requirements.txt
langchain>=0.2.11
langgraph>=0.2.14
langchain-openai>=0.1.22
openai>=1.44.0
faiss-cpu>=1.8.0
pymongo>=4.8.0
mcp>=1.1.0
typer>=0.12.3
pydantic>=2.8.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
rich>=13.8.0
numpy>=1.26.4
